requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
lxml==5.4.0
signxml==4.2.0
cryptography==43.0.3
